Create utility classes for handling encryption/decryption
Summary:
These classes are convenient to use for encrypt/decrypt sensitive data in Android.

The implementation is using algorithm RSA/AES to encrypt or decrypt data.

The implementation has 2 different ways because the class **KeyGenParameterSpec** is not available in android version below 6.0. https://developer.android.com/reference/android/security/keystore/KeyGenParameterSpec.html

1. For Android before version 6.0 (Pre-Marshmallow)

2. For Android after version 6.0 (Marshmallow)

OMGKeyManager will check the API version to apply the appropriate implementation.

Test Plan:
It cannot test with local unit test na ja.

The implementation is relate with android resources.

It should be android instrumental test which is not available in android library module.

I think we should implement this test in the app module na where is the only place to testtt.

Otherwise, we test it in the sample app which I think it may not appropriate :(

The app module is here https://phabricator.omisego.io/diffusion/AND/browse/master/app/

p.s. Actually, I've already tried to export the library to test the functionality in the sample app and it's work very well na

Reviewers: thibault, mederic, sirn

Reviewed By: thibault, mederic, sirn

Maniphest Tasks: T458

Differential Revision: https://phabricator.omisego.io/D263